Updated database from LocalDb to DESKTOP-7IUSC5H.
Used commands:
1) Enable-Migrations -EnableAutomaticMigrations -Force
2) Add-Migration Initial
3)update-database

Followed steps:

Delete existing migrations from Migrations_History table.

Delete existing migrations from the Migrations Folder.

Run add-migration Reset. This will create a migration in your Migration folder that includes creating the tables (but it will not run it so it will not error out.)

You now need to create the initial row in the MigrationHistory table so EF has a snapshot of the current state. EF will do this if you apply a migration. However, you can't apply the migration that you just made as the tables already exist in your database. So go into the Migration and comment out all the code inside the "Up" method.

Now run update-database. It will apply the Migration (while not actually changing the database) and create a snapshot row in MigrationHistory.

You have now reset your migrations and may continue with normal migrations.